requirements.txt
asgiref
Django
psycopg2-binary
sqlparse
tzdata
whitenoise
environs
djangorestframework
Pillow
drf-spectacular
ruff
djangorestframework-simplejwt
djoser